// ACP-Checkout-Gateway/src/api.ts
export interface components {
    schemas: {
        Address: {
            name: string;
            line_one: string;
            line_two?: string;
            city: string;
            state: string;
            country: string;
            postal_code: string;
        };
// ...
        FulfillmentDetails: {
            name?: string;
            phone_number?: string;
            /** Format: email */
            email?: string;
            address?: components["schemas"]["Address"];
        };
        Buyer: {
            first_name: string;
            last_name: string;
            /** Format: email */
            email: string;
            phone_number?: string;
        };
        Item: {
            id: string;
            quantity: number;
        };
// ...
        LineItem: {
            id: string;
            item: components["schemas"]["Item"];
            base_amount: number;
            discount: number;
            subtotal: number;
            tax: number;
            total: number;
            name?: string;
            description?: string;
            images?: string[];
            unit_amount?: number;
            disclosures?: components["schemas"]["Disclosure"][];
            custom_attributes?: components["schemas"]["CustomAttribute"][];
            marketplace_seller_details?: components["schemas"]["MarketplaceSellerDetails"];
        };
        Total: {
            /** @enum {string} */
            type: "items_base_amount" | "items_discount" | "subtotal" | "discount" | "fulfillment" | "tax" | "fee" | "total";
            display_text: string;
            amount: number;
            description?: string;
        };
        FulfillmentOptionShipping: {
            /** @constant */
            type: "shipping";
            id: string;
            title: string;
            subtitle?: string;
            carrier?: string;
            /** Format: date-time */
            earliest_delivery_time?: string;
            /** Format: date-time */
            latest_delivery_time?: string;
            subtotal: number;
            tax: number;
            total: number;
        };
// ...
        MessageInfo: {
            /** @constant */
            type: "info";
            /** @description RFC 9535 JSONPath */
            param?: string;
            /** @enum {string} */
            content_type: "plain" | "markdown";
            content: string;
        };
// ...
        Order: {
            id: string;
            checkout_session_id: string;
            /** Format: uri */
            permalink_url: string;
        };
// ...
        CheckoutSession: components["schemas"]["CheckoutSessionBase"];
// ...
    };
    responses: never;
    parameters: {
        Authorization: string;
        ContentType: string;
        AcceptLanguage: string;
        UserAgent: string;
        IdempotencyKey: string;
        RequestId: string;
        Signature: string;
        Timestamp: string;
        APIVersion: string;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}